Move bigIntToBytes to the utils (dlog) package
I need to use it in the crypto package
